backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
pydantic
python-dotenv
